ci(docker-tests): split SLP tests into separate feature-flagged job
- Add `docker-tests-slp` feature flag to mm2_main/Cargo.toml
- Guard slp_tests module with #[cfg(feature = "docker-tests-slp")]
- Move SLP-specific helpers from docker_tests_common.rs to slp_tests.rs
- Add docker-tests-slp CI job that only starts FORSLP container
